Factor out the timer clock into its own Composable
State travels down the hierarchy, events bubble up. Here, the `Timer` is
getting it's currently displayed duration as state from the containing
`App`. Through a `LaunchEffect`, it generates an event after some time
has elapsed; it is the responsibility of the event handler to update
the duration accordingly. That way, the duration state is hoisted out
of the (now stateless) `Timer` component and the `App` can implement
timer reset all by itself and can use the same state with other
components.