Switch to a list of interval for state
The previous approach was prone to a "drift" of the timer compared to
the actual clock because we would time individual durations in the
`LaunchedEffect` coroutine and just add them up.

The new approach does have this problem; the coroutine is also no longer
cancelled every tick. Instead the `Timer` now doesn't depend on any
state itself, other than whether it's running. It updates the time
interval through events. The intervals are then aggregated in a list in
such a way that overlapping intervals are merged.

The new state list also provides the richer information that is
necessary to display the individual segments of work in a list.